[#140] Don't distinguish between happenings booked/verified by user

+++ resources/js/Composables/ModalActions.js
@@ -2,7 +2,7 @@ import HappeningModal from "@/Components/Modals/HappeningModal.vue";
 import ResourceInfoModal from "@/Components/Modals/ResourceInfoModal.vue";
 import { useHappeningStore } from "@/Stores/HappeningStore";
 import useModal from "@/Stores/Modal";
-import {trans} from "laravel-vue-i18n";
+import { trans } from "laravel-vue-i18n";
 
 export function useCreateModal(happening) {
     const happeningStore = useHappeningStore();
@@ -10,13 +10,13 @@ export function useCreateModal(happening) {
     return {
         view: HappeningModal,
         content: {
-            title: trans('modal.create.title'),
-            description: trans('modal.create.description'),
+            title: trans("modal.create.title"),
+            description: trans("modal.create.description"),
         },
         payload: { ...happening, editable: true },
         actions: [
             {
-                label: trans('modal.create.action.create'),
+                label: trans("modal.create.action.create"),
                 callback: (happening) => {
                     return happeningStore.addHappening(happening);
                 },
@@ -31,13 +31,13 @@ export function useInfoModal(happening) {
     return {
         view: HappeningModal,
         content: {
-            title: trans('modal.info.title'),
-            description: trans('modal.info.description'),
+            title: trans("modal.info.title"),
+            description: trans("modal.info.description"),
         },
         payload: { ...happening, editable: false },
         actions: [
             {
-                label: trans('modal.info.action.ok'),
+                label: trans("modal.info.action.ok"),
                 callback: () => {
                     modal.close();
                 },
@@ -46,24 +46,45 @@ export function useInfoModal(happening) {
     };
 }
 
-export function useVerifyModal(happening) {
+export function useVerifyModalFromPermissions(happening, can = happening.can) {
     const happeningStore = useHappeningStore();
+
+    const actions = [];
+    if (can.verify) {
+        actions.push({
+            label: trans("modal.verify.action.verify"),
+            callback: (happening) => {
+                return happeningStore.verifyHappening(happening);
+            },
+        });
+    }
+
+    if (can.edit) {
+        actions.push({
+            label: trans("modal.edit.action.update"),
+            callback: (happening) => {
+                return happeningStore.editHappening(happening);
+            },
+        });
+    }
+
+    if (can.delete) {
+        actions.push({
+            label: trans("modal.delete.action.delete"),
+            callback: (happening) => {
+                return happeningStore.deleteHappening(happening.id);
+            },
+        });
+    }
 
     return {
         view: HappeningModal,
         content: {
-            title: trans('modal.verify.title'),
-            description: trans('modal.verify.description'),
+            title: trans("modal.verify.title"),
+            description: trans("modal.verify.description"),
         },
-        payload: { ...happening, editable: false },
-        actions: [
-            {
-                label: trans('modal.verify.action.verify'),
-                callback: (happening) => {
-                    return happeningStore.verifyHappening(happening);
-                },
-            },
-        ],
+        payload: { ...happening, editable: can.edit },
+        actions,
     };
 }
 
@@ -73,13 +94,13 @@ export function useEditModal(happening) {
     return {
         view: HappeningModal,
         content: {
-            title: trans('modal.edit.title'),
-            description: trans('modal.edit.description'),
+            title: trans("modal.edit.title"),
+            description: trans("modal.edit.description"),
         },
         payload: { ...happening, editable: true },
         actions: [
             {
-                label: trans('modal.edit.action.update'),
+                label: trans("modal.edit.action.update"),
                 callback: (happening) => {
                     return happeningStore.editHappening(happening);
                 },
@@ -94,13 +115,13 @@ export function useDeleteModal(happening) {
     return {
         view: HappeningModal,
         content: {
-            title: trans('modal.delete.title'),
-            description: trans('modal.delete.description'),
+            title: trans("modal.delete.title"),
+            description: trans("modal.delete.description"),
         },
         payload: { ...happening, editable: false },
         actions: [
             {
-                label: trans('modal.delete.action.delete'),
+                label: trans("modal.delete.action.delete"),
                 callback: (happening) => {
                     return happeningStore.deleteHappening(happening.id);
                 },
@@ -109,30 +130,36 @@ export function useDeleteModal(happening) {
     };
 }
 
-export function useEditDeleteModal(happening) {
+export function useEditModalFromPermissions(happening, can = happening.can) {
     const happeningStore = useHappeningStore();
+
+    const actions = [];
+    if (can.edit) {
+        actions.push({
+            label: trans("modal.edit.action.update"),
+            callback: (happening) => {
+                return happeningStore.editHappening(happening);
+            },
+        });
+    }
+
+    if (can.delete) {
+        actions.push({
+            label: trans("modal.delete.action.delete"),
+            callback: (happening) => {
+                return happeningStore.deleteHappening(happening.id);
+            },
+        });
+    }
 
     return {
         view: HappeningModal,
         content: {
-            title: trans('modal.edit_delete.title'),
-            description: trans('modal.edit_delete.description'),
+            title: trans("modal.edit_delete.title"),
+            description: trans("modal.edit_delete.description"),
         },
-        payload: { ...happening, editable: true },
-        actions: [
-            {
-                label: trans('modal.edit_delete.action.update'),
-                callback: (happening) => {
-                    return happeningStore.editHappening(happening);
-                },
-            },
-            {
-                label: trans('modal.edit_delete.action.delete'),
-                callback: (happening) => {
-                    return happeningStore.deleteHappening(happening.id);
-                },
-            },
-        ],
+        payload: { ...happening, editable: can.edit },
+        actions,
     };
 }
 
@@ -142,13 +169,13 @@ export function useResourceInfoModal(resourceInfo) {
     return {
         view: ResourceInfoModal,
         content: {
-            title: trans('modal.resource_info.title', {resource_title: resourceInfo.resource.title}),
-            description: trans('modal.resource_info.description'),
+            title: trans("modal.resource_info.title", { resource_title: resourceInfo.resource.title }),
+            description: trans("modal.resource_info.description"),
         },
         payload: { ...resourceInfo },
         actions: [
             {
-                label: trans('modal.resource_info.action.ok'),
+                label: trans("modal.resource_info.action.ok"),
                 callback: () => {
                     modal.close();
                 },
